Fix ServiceInstantiation_test_xx
Fix the last handle set for characteristics
Fix discovery order.
From a user perspective, the service tree is traversed in deep.

The previous behavior was:

for service in services:
    service_callback(callback)

    for characteristic in characteristics:
        characteristic_callback(characteristic)

The new behavior is:

for service in services:
	service_callback(callback)
	        for characteristic in service.characteristics:
		        characteristic_callback(characteristic)

Reserve the valid count of attribute records when declaring a service.
Report error when it is not possible to add a service.