Fix GattClient read and possible overide of current gatt operation.
Instead of calling the GattClient read callback once the data is received, the
callback is called once the procedure has completed, this prevent users to
launch another procedure while another one is already ongoing.

It was also possible that the flag currentState in the function
gattProcedureCompleteCB got overriden by a user callback. Reordering the
reset of this flag and returning immediatelly once the treatment has been
done prevent this wrong behavior.

As a consequence, it is not required for a client to wait before calling
aci_gatt_read_charac_val.