Fix forwarding of isServiceDiscoveryActive and terminateServiceDiscovery to clients.
The previous code was just forwarding the function to the first _numClient. This
would work if clients the connection pool are accessible sequentially unfortunatelly
it is not the case.

With this fix the entirety of the connection pool is traversed and if a client is
present then the action is forwarded.